lkl: rework syscall implementation
Rework the syscall implementation to remove the dependency on "with
data" interrupts. In the new implementation lkl_syscall() saves the
syscall information directly into variables accesible from Linux
threads. Because of that we need to use GCC's built-in functions for
atomic memory access to synchronize access between Linux and non-Linux
threads.

Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>