lkl tools: virtio: prepare for network device support
Change the queue callback and add a return value to allow the device to
stop the queue processing and add a virtio_dev parameter to the
check_features callback.

The queue processing implementation is also simplfied and the virtio
request is now holding a limited number of buffers which should be
enough for most devices. This allows us to remove the dynamic allocation
for requests since it is impossible handle allocation failures at this
level.  If needed, we can later add new APIs to allow the device to get
buffers past the limit imposed in the virtio request structure.

Finally, a few cosmetic changes are made to function and callback names.

Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>