lkl: rework irq handling
Now that we don't need to support "with data" interrupts we can signal
interrupts without the need of using host lock / semaphores. This
removes an important limitation of the implementation: before this patch
we were not able to trigger an interrupt from a softirq or interrupt
handler.

The new implementation uses GCC's built-in functions for atomic memory
access to synchronize and simple boolean flags.

Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>